complete backend

+++ backend/controllers/product.controller.js
@@ -60,17 +60,21 @@ export const createProduct = async (req, res) => {
   }
 };
 
-export const  deleteProduct = async (req, res) => {
-    const { id } = req.params;
-    console.log("id:", id);
-  
-    try {
-      await Product.findByIdAndDelete(id);
-      res.status(200).json({ success: true, message: "Product Delete" });
-    } catch (error) {
-      res.status(404).json({
-        success: false,
-        message: "Product Not found",
-      });
-    }
+export const deleteProduct = async (req, res) => {
+  const { id } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Invalid Product Id" });
   }
+
+  try {
+    await Product.findByIdAndDelete(id);
+    res.status(200).json({ success: true, message: "Product Delete" });
+  } catch (error) {
+    res.status(500).json({
+      success: false,
+      message: "Server Error",
+    });
+  }
+};
